requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.1.0
beautifulsoup4==4.9.1
chardet==3.0.4
discord==1.0.1
discord.py==1.4.1
google==3.0.0
idna==2.10
multidict==4.7.6
python-dotenv==0.14.0
soupsieve==2.0.1
typing-extensions==3.7.4.3
yarl==1.5.1
